code/run_templates/consume_to_api/engine/requirements.txt
azure-identity
cosmotech-api-python~=3.0.0
